Eliminated the output curves as I've realized they were pointless. This causes the generator to now produce the correct amount of resource at all times, only input resource is managed by a curve .
Currently, and as best as I can tell previously, partially fulfilled input resource requests will not impact generator performance until any given resource drops below 80% of the request, then it will flame out. For That Frame, all resources will be consumed, outputs will not be produced, and the generator will then disable itself

OnUpdate has been cut into a bunch of different methods, this is largely pointless and I've done it while experiencing worrying side effects of energy drinks, an intense headache, severe confusion as to what was going on, probably should have seeked medical attention.

Update the input curve of the generator part, and reduced it's output considerably Generator now runs at 50-70% throttle for test rover, medium sized manned 4 wheels. Updated curve attempts to add power bands where the generator is more efficient.

Currently there is no throttle overruning, with the removal of the curves it does not appear to be necessary, the code is commented out in case it is necessary